requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
requests==2.33.1
python-dotenv==1.2.2
